Update logic moved to MainActivity + WIP Test for In-App update
